feat(newrelic): APM application id can be looked up by name
This is intended to be a quality of life addition to allow
`ApplicationSet` resources to minimize boilerplate when dealing with
NewRelic APM resources across multiple environments.
- An application programmatically [uses the NR SDK to report to
  NewRelic](https://github.com/newrelic/go-agent/blob/master/v3/examples/server/main.go#L267-L274)
- NewRelic APM Application ID has to then be retrieved following the
  first run of an application from the NR Dashboard and added to an
  ArgoCD `Application` via the annotation, across multiple environments.
Using ArgoCD `ApplicationSet` resources, the annotation can be either
provided via app_id, or by name, e.g. `MyApp-{{ .values.env }}`
If `dest.Recipient` can be parsed to an int, then app_id is provided
and logic remains as before.
If `dest.Recipient` cannot be parsed to an integer, then it was passed
by name and we call `/v2/applications.json` to query using
`filter[name]`
If `/v2/applications.json` returns multiple application IDs then an
error is returned, as we can't determine which app_id to use to place
the deployment marker.

Signed-off-by: 3bbbeau <0x3bb@3bb.io>