feat: add ability to use custom certs for email
With argoproj/argo-cd#15394 it has been possible to use custom certs from
`argocd-tls-certs-cm`. This works as intended for any notification that
is based on HTTP, but _not_ for emails (since the email doesn't make use
of `http.Transport`).

This commit fixes that by extracting out the TLS config to a separate
function which can be called by the email service. The functionality
which creates an email has been extracted from the library which was
previously used since there's no support for setting custom TLS configs.

This also means we can get rid of the `gomodules.xyz/notify` library.

Signed-off-by: Blake Pettersson <blake.pettersson@gmail.com>